final commit, now the app is fully functional

+++ todo/todo.py
@@ -12,7 +12,9 @@ bp = Blueprint('todo', __name__)
 def index():
     db, c = get_db()
     c.execute(
-        'select t.id, t.description, u.username, t.completed, t.created_at from todo t JOIN user u on t.created_by = u.id order by created_at desc;'
+        'select t.id, t.description, u.username, t.completed, t.created_at'
+        ' from todo t JOIN user u on t.created_by = u.id where t.created_by = %s order by created_at desc',
+        (g.user['id'],)
     )
      
     todos = c.fetchall()
@@ -81,8 +83,8 @@ def update(id):
             db, c = get_db()
             c.execute(
                 'update todo set description = %s, completed = %s'
-                ' where id = %s',
-                (description, completed, id)
+                ' where id = %s and created_by = %s',
+                (description, completed, id, g.user['id'])
             )
             db.commit()
             return redirect(url_for('todo.index'))
@@ -94,6 +96,6 @@ def update(id):
 @login_required
 def delete(id):
     db, c = get_db()
-    c.execute('delete from todo where id = %s', (id,))
+    c.execute('delete from todo where id = %s and created_by = %s', (id, g.user['id']))
     db.commit()
     return redirect(url_for('todo.index'))
